# optimizing ordinal_encoding