I've added an optimization to `OrdinalEncoder.ordinal_encoding`. When the number of categories is large, `OrdinalEncoder` (and also other encoders using `OrdinalEncoder`) is very slow because of the for loops and `pandas.DataFrame.iloc`. So, for loops in `ordinal_encoding` are removed and replaced by using `pandas.Series.map`. 

All `nosetests --with-coverage` tests has been passed.

Any comment is warmly welcome. Here is my benchmarks.

```py
# on master branch 
In [1]: import category_encoders as ce

In [2]: %%time
   ...: many_categories_list = ["category_{}".format(i) for i in range(10000)]
   ...: encoder = ce.OrdinalEncoder()
   ...: encoder.fit(many_categories_list)
   ...: encoder.transform(many_categories_list)
   ...:
CPU times: user 1min 8s, sys: 835 ms, total: 1min 9s
Wall time: 1min 11s
```

```python
# on this branch
In [1]: import category_encoders as pr

In [2]: %%time
   ...: many_categories_list = ["category_{}".format(i) for i in range(10000)]
   ...: encoder = pr.OrdinalEncoder()
   ...: encoder.fit(many_categories_list)
   ...: encoder.transform(many_categories_list)
   ...:
CPU times: user 31.9 ms, sys: 3.47 ms, total: 35.4 ms
Wall time: 34.3 ms
```