@wdm0006 2018-02-09
Wow that's a pretty significant speed up, thanks for the PR, this looks great.